Move handlers into its own module
The v0 handlers are moved to v0_handlers.rs to separate them
from the future v1 message handlers.
